rideshare2/requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
mysql-connector-python>=8.0.0
